+++ src/main/java/com/project/EmpowerHR/Repositories/JobDepartementRepository.java
@@ -4,4 +4,4 @@ import com.project.EmpowerHR.Entities.JobDepartmentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JobDepartementRepository extends JpaRepository
-        <JobDepartmentEntity, Integer> {}
+        <JobDepartmentEntity, Long> {}
