+++ src/main/java/com/project/EmpowerHR/Repositories/LeaveRepository.java
@@ -4,4 +4,4 @@ import com.project.EmpowerHR.Entities.LeaveEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LeaveRepository extends JpaRepository
-        <LeaveEntity, Integer> {}
+        <LeaveEntity, Long> {}
