+++ src/main/java/com/project/EmpowerHR/Repositories/PayrollRepository.java
@@ -4,4 +4,4 @@ import com.project.EmpowerHR.Entities.PayrollEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PayrollRepository  extends JpaRepository
-        <PayrollEntity, Integer> {}
+        <PayrollEntity, Long> {}
