+++ src/main/java/com/project/EmpowerHR/Repositories/SalaryRepository.java
@@ -4,4 +4,4 @@ import com.project.EmpowerHR.Entities.SalaryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SalaryRepository  extends JpaRepository
-        <SalaryEntity, Integer> {}
+        <SalaryEntity, Long> {}
